esphome_mqtt_to_influx/requirements.txt
influxdb-client
paho-mqtt
click


